test(entity): split type tests into .test-d.ts and hybrid .spec.ts files
Move pure type-only tests (using expectTypeOf) to .test-d.ts files so
they are only compiled and never executed. Keep hybrid tests that use
@ts-expect-error in .types.spec.ts files.
